finalizando nova estruturação de login
