+++ src/components/Cadastrar/style.css
@@ -1,0 +1,10 @@
+/* ------------------------------------Formulário de cadastro----------------------------------- */
+.signUp{
+    width: 100vw;
+    margin: 5vh 5vw;
+    display: inline-block;
+}
+.btn-cadastrar{
+    margin: 10px 5px 10px;
+    height: 40px ;
+}
