--- src/components/MainMinhaConta/style.css
@@ -1,74 +1,0 @@
-:root {
-    --color-secondary: #39ff14;
-    --color-dark: #343a40;
-  }
-/* ----------------------------------------------------BODY----------------------------------- */
-.formularios {
-   display: flex;
-   flex-flow: row nowrap;
-   justify-content: center;
-  }
-  
-  h1,
-  h4 {
-    display: flex;
-    justify-content: center;
-    text-align: center;
-    align-items: center;
-  }
-
-/* ------------------------------------Formulário de entrada----------------------------------- */
-
-.signIn{
-    width: 45vh;
-    margin: 20px;
-    display: inline-block;
-}
-.bodyForm{
-    background-color: #EEEEEE;
-    border-bottom-left-radius: 10% ;
-    border-bottom-right-radius: 10% ;
-    padding: 10px;
-}
-
-.form-control{
-    background-color: #c4c4c4;
-}
-
-.headerForm{
-    background-color: black;
-    color: white;
-    border-top-left-radius: 15% ;
-    border-top-right-radius: 15% ;
-    padding: 10px;
-}
- .esqueci{
-     display: inline-block;
-     text-decoration: underline ;
-     padding: 0px 20px 10px 0px;
- }
-
- .esqueci a{
-     color: initial;
- }
-
- .btn-entrar{
-     display: block;
-     margin: 10px 0px;
- }
- .btn-entrarCom{
-    display: inline-block;
-    margin:20px 10px;
-}
- 
-/* ------------------------------------Formulário de cadastro----------------------------------- */
-.signUp{
-    width: 45vh;
-    margin: 20px;
-    display: inline-block;
-}
-.btn-cadastrar{
-    margin:30px 0px 0px 50px;
-    height: 40px ;
-
-}
