+++ src/pages/Cadastro/index.js
@@ -1,0 +1,15 @@
+import Cadastrar from '../../components/Cadastrar'
+
+const Cadastro = () => {
+    return (
+        <>
+            <div className="titleForm">
+                <h1>Cadastre-se</h1>
+                <h4>Faça seu cadastro para continuar.</h4>
+            </div>
+            <Cadastrar />
+        </>
+    )
+}
+
+export default Cadastro
