+++ src/pages/Login/index.js
@@ -1,0 +1,14 @@
+import Entrar from "../../components/Entrar";
+
+const MinhaConta = () => {
+  return (
+    <>
+        <div className="titleForm">
+          <h1>Minha Conta</h1>
+          <h4>Você precisa ser cadastrado ou já ter um cadastro para continuar.</h4>
+        </div>
+        <Entrar />
+    </>
+  );
+};
+export default MinhaConta;
