--- src/pages/MinhaConta/Reenvio/esqueciLogin.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="reenvio.css">
-    <title>Esqueci meu login</title>
-</head>
-
-<body>
-    <form name="reenvio">
-            <table>
-                <tbody>
-                    <tr>
-                        <td className="headerTitulo">
-                            <p className="tituloForm">
-                                Esqueceu seus dados para acesso?</br>
-                                Nós o renviaremos para seu e-mail</font>                         
-                            </p>
-                        </td>
-                    </tr>
-                    <tr>
-                        <td className='descricaoForm'>
-                            <p>Se você esqueceu sua senha para acesso
-                                ao site, não se preocupe. Apenas informe no campo a seguir o
-                                número do CPF ou CNPJ de cadastro que você possui no site e
-                                reenviaremos seus dados de acesso para o e-mail cadastrado.</p><br>
-                            <p>CPF ou CNPJ<br>
-                                <input name="cpf_cnpj" type="text" className="formulario_1" size="50" tabindex="1">
-                            </p>
-                            <p><input src="../../../assets/img/btn_enviar.png"name="I1"  type="image"></p>
-                        </td>
-                    </tr>
-                </tbody>
-            </table>
-        <input type="hidden" name="funcao" value="reenviar">
-    </form>
-</body>
-
-</html>
