--- src/pages/MinhaConta/Reenvio/reenvio.css
@@ -1,27 +1,0 @@
-body {
-    background-color: #fff;
-    margin: 0;
-}
-
-.headerTitulo {
-    background-color: #343a40;
-    background-position: bottom; 
-    padding-left:14px; 
-    padding-top:22px;
-    border-radius: 20px 20px 0px 0px;
-}
-
-.tituloForm {
-    font-size: 18px;
-    color: #fff;
-}
-
-.descricaoForm {
-    height: 200px;
-    vertical-align: top;
-    padding:14px
-}
-
-.formulario_1 {
-    size: 60px;
-}
