--- src/pages/MinhaConta/index.js
@@ -1,10 +1,0 @@
-import MainMinhaConta from "../../components/MainMinhaConta";
-
-const MinhaConta = () => {
-  return (
-    <>
-      <MainMinhaConta />
-    </>
-  );
-};
-export default MinhaConta;
