+++ src/routes/index.js
@@ -3,7 +3,9 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Home from "../pages/Home";
 import Sobre from "../pages/Sobre";
 import Contato from "../pages/Contato";
-import MinhaConta from "../pages/MinhaConta";
+import Login from "../pages/Login";
+import Cadastro from '../pages/Cadastro'
+import Senha from '../pages/Login/ReenvioSenha'
 
 const Routes = () => {
   return (
@@ -11,7 +13,9 @@ const Routes = () => {
       <Switch>
         <Route path="/contato" component={Contato} />
         <Route path="/sobre" component={Sobre} />
-        <Route path="/minhaConta" component={MinhaConta} />
+        <Route path="/login" component={Login} />
+        <Route path='/esqueci-senha' component={Senha} />
+        <Route path="/cadastro" component={Cadastro} />
         <Route path="/" component={Home} />
       </Switch>
     </Router>
